BUG: Committing for Dean Inglis.
The widget's BuildRepresentation() serves to take its ivar vtkPlaneSource and generate a 2D polyline outline of its bounds.

This should only be called after UpdatePlane since within that method, RestrictPlaneToVolume is enforced, the vtkPlaneSource could be modified, and hence, the outline rep will be inconsistent.

This manifests itself when interacting as though the plane outline "bounces" off the widget's resliced image and then bounces back as one tries to drag past the input image data bounds.