ENH: Generic vtkInformation Has and Remove methods
Most of the vtkInformationKey subclasses had identical "Has" methods, so
we replace them with one method in the superclass (as was already the
case for "Remove").  This also allows creation of generic
vtkInformation::Has and vtkInformation::Remove overloads for the
abstract vtkInformationKey type.

The vtkInformationKey::Has and vtkInformationKey::Remove methods are
virtual for safe handling of the Request key optimization.  Fortunately
in most cases we can bypass the vtable when the key type is known.