ENH: Improvements to TriggerRMI()
For small messages, vtkMultiProcessController::TriggerRMI() sends
the message as a single packet (instead of two as was the
case before). Two packets are still sent when the message size is greater
than (128-4) ints.

The length specified in vtkCommunicator::Receive() is treated as the max-length
and error is raised only when the max-length is less than the message length
(similar to MPI_Recv). To get the actual length of the received message, added a
new GetCount() which returns the number of items (as against number of bytes)
read in by the last receive.

Cleaned up code in vtkMPICommunicator that adds support for data-sizes larger
than VTK_INT_MAX.