ENH: Attempting to solve z-fighting for surface-with-edges by offsetting polygons instead of lines. This seems to work better than offsetting lines. However offsetting polygons produces some artifactes at boundaries. Hence we use the offsetting only when rendering surface with edges and not otherwise.
ENH: Reverting the order of painters in vtkDefaultPainter particularly to ensure
that the display list painter is before the vtkRepresentationPainter. This is
essential since we cannot reuse the display list for rendering surfaces and
edges consequently, we were ending up regenerating the display lists on each
render.

ENH: ParaView no longer enable any coincident topology resolution technique.
This is not needed since surface with edges handles coincident topology on its
own.