ENH: Adding vtkApplyIcons and support for attribute types.
vtkApplyIcons, much like vtkApplyColors, defines the
icon indices for elements of a dataset through a lookup table,
default icon, annotations, and/or the current selection.

There are a number of places that specify attribute types
(point, cell, vertex, edge, row, field) and create an enumeration
for this (vtkDataObjectToTable, vtkSelectionNode, ...). A new
method GetAttributes(int) in vtkDataObject uses an enum for the
possible attribute types and subclasses retrieve the appropriate
attributes. GetAttributesAsFieldData(int) performs the same function
but returns a vtkFieldData instance instead of vtkDataSetAttributes.
GetAttributeTypeForArray is another convenience method that returns
the attribute type that an array is contained in.