ENH: Initial import of the work 2D API, 2D scene and charts code.
This is an initial import of the surviving parts of the 2D drawing API I
have been working on, along with a prototype for a 2D scene and some chart
elements. This work is experimental, and parts of the API will be moved into
Rendering.

The API is subject to change and/or removal.