COMP: This should fix a visibility issue on Apple with GCC 4.2.
This one was very tricky, thanks to Brad King for tracking this down. If
vtkStdString is not exported then with Apple's GCC 4.2 the symbols for
vtkArrayIteratorTemplate<vtkStdString> are not exported. This may need to be
turned into a macro someday, but so far this appears to be the only instance.